Convert some .nibs to a newer format.
The original .nib interface files were in the original NeXTStep format,
which is not compatible with modern versions of OS X. Convert these to
a slightly more modern format that OS X can read.

This is actually still an old format by modern standards. Conversion was
performed by loading and saving the .nibs in the Rhapsody DR2 version of
Interface Builder, which was simultaneously still able to read NeXTStep
.nibs while using a newer format for writes. Modern OS X uses the
XML-based .xib format instead, but at least for the time being they
can be opened and edited.

Not all the .nibs were successfully translated. A few failed to
translate properly for various reasons (some used an old text control
class that no longer exists, others just read as corrupt). These will
have to be fixed later or recreated manually.

While we're doing this, remove Project.nib~ which is an old backup
copy of Project.nib that was included in the source dump.