Add stub Display PostScript API.
NeXTStep used Display PostScript to render the screen, but this API
no longer exists in OS X. There is a very similar API called Quartz
which can probably be used instead. This new file adds empty stub
functions for all the functions in the DPS API; the idea is to
partially implement the API using Quartz as the backend.