Add compatible "Storage" class.
Lots of the DoomEd code uses a class called "Storage" which was in the
NeXTStep API and implemented an array for C base types. This no longer
exists in modern Objective C / Cocoa. Add a basic reimplementation of
it and switch over code that uses "Storage" to the new
"CompatibleStorage" implementation.